requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.9
pydantic-settings==2.10.1
google-cloud-bigquery==3.38.0
google-cloud-firestore==2.21.0
google-auth==2.40.3
